if in start

+++ start.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-sudo apt-get install -y curl ca-certificates brotli
-curl https://mmdbcdn.posthog.net/ | brotli -d > mmdb.db
+if [[ ! -f mmdb.db ]]; then
+    sudo apt-get install -y curl ca-certificates brotli
+    curl https://mmdbcdn.posthog.net/ | brotli -d > mmdb.db
+fi
+
 ./livestream
